Significant Changes Made the following changes to the
LexicalLookup class (these changes are noteworthy although they do not require changes to any existing code and in fact, the enhance it in various ways.

I: references to Word instances can no longer be looked up unless they are statically known to be Adjective, Adverb, Noun, or Verb instances. Previously any word could be looked up dynamically. But if for example, that word was a preposition, a type with no WordNet Lookup, a NoSynonymLookupForTypeException would be thrown. These Exceptions are now avoided by the requirement that the static reference be of the correct type. The previous unsafe, dynamic functionality was no being used by any code, and I felt that the presence of a Lookup overload taking an instance of the base class and throwing an exception was confusing and hard to predict.


II: New functionality, with no cost or learning curve has been added. 
Here is how it works, the various overloarded LexicalLookup functions: IsSynonymFor and IsSimilarTo used to return bool results.
However, while this was very convenient when you wanted a simple yes or no answer, 
it meant you had to call another method, and incur additional overhead, 
if you also needed to know the specific numeric Similarity Ratio between the elements being compared by the functions.

Now you can have your cake and eat it!

IsSynonymFor and IsSimilarTo now return instances of a struct type ,containing a single bool and a single double which package both the true/false result and the ratio.
However, instances of the struct implicitly convert to the appropriate value in any applicable context. 
so if you want to compare the result of IsSymilarTo directly to a numeric value, it will just work as it converts to double in such a context.
similarly, if you want to control the flow of an if else block with the result of IsSymilarTo, it will just work as it converts to bool in that context.

Full details on the struct are provided below.
Read them because it will make me happy!



This is functionality is provided via the SimResult struct described below.

Encapsulates multiple pieces of information gathered during a similarity comparison into a light weight type.
The structure cannot be created from outside of the LexicalLookup class and is used to convey internal results.
No special syntax is or code changes are required to manipulate this type. It will implicitly convert to bool
So all code with forms such as:  if ( a.IsSimilarTo(b) ) { ... } need not and should not be changed. 
However, If the numeric ratio used to determine similarity is needed and applicable, the type will implicitly convert
to a double. This removes the need for public code such as: if ( LexicalLookup.GetSimiliarityRatio(a, b) > 0.7 ) { ... }
Instead one may simply write the same logic as if: if ( a.IsSimilarTo(be) > 0.7 ) { ... }

The other reason for the introduction of this type is that, as it also is now the default return type for IsSynonymFor invocations,
it allows the results provided by such calls to provide additional information as we learn how to glean the wealth of the WordNet data.

git-svn-id: http://vps6982.inmotionhosting.com/svn/lasi@828 49725315-0998-4498-bb4a-373bc1a52cb2