The UI now has a functioning options menu: Under Tools -> Preferences from either the ProjectPreviewScreen or the ResultsScreen. So far, the options are: Automatically name new projects, Minimize to the system tray, PerforamanceLevel (High, Normal, or Low)
Program options now load automatically on startup and save automatically when changed.

Minimizing to the system tray is now implemented. (A message will be displayed in the tray if the program is in the system tray minimized state when analysis completes.
All windows may be minimized to the system tray except the StartupScreen.

MISC:
Fixed an issue which allowed more documents per project than intended.
Added a little documentation.


git-svn-id: http://vps6982.inmotionhosting.com/svn/lasi@840 49725315-0998-4498-bb4a-373bc1a52cb2