Improved and Generalized Control Flow
The TypedSwitch structure has been replaced with a much more powerful and generalizable pattern matching syntax. In addition to being more expressive, it allows for 'case style' expressitions to return values as like any other function call.
Began building further on this with an algebraic type decomposer (already almost complete) which will allow for the destructuring of types and the application of corresponding functions to each one.