Fixed some issues. Fixed a VerbConjugation issue which caused a null reference exception to be caught at an inappropriate teir, thus resulting in empty synonym results for certain words. Fixed issue in Phrase.ToString(), when verbose mode was enabled, which resulted in infinite mutually recursive calls to ToString between bound Phrases.
As an aside, results exposed to the user continue to improve.
Classifiers and Posessors (is a, has a) are now working correctly thanks to the Conjugation issue fix mentioned above. Performance also continues to remain excellent, even as more and more bindings are taking place thanks to improvements (still have a long way to go) in Object Binding.