Refactored a large amount of the UI.
Moved duplicated/semi duplicated methods to an internal static class.
Made drag and drop behavior consistent across all windows.
Dragging and dropping a document of a supported type will add it to the project on all screens except the InProgressScreen.
Dragging a folder will recursively add all files within it which are of a supported type.
Improved consistency of alert messages displayed when the user tries to add an invalid or already added document, or when the project is full.

Max documents increased to 10.