Lots of refactoring
Reorganized Algorithm namespaces
Reduced code duplication in WordNet lookups
Renamed and cleaned up many classes and methods
The IsClassifer and IPossessive methods of the Verb and VerbPhrase
classes no longer overlap: That is an instance may be either a
Classification indicator ( loose is a relationship ) or a Poessession
indicator ( loose has a relationship ) or neither, but no longer both.
Improved overall performance and internal organization of the Algorithm
assembly. Discovered an issue in the Adverb and Adjective Lookup classes
that involves references to other sets, need to look at this in detail.