Resolved the conflict between the Nav bar and the Element context menus. I had bound an event handler to fire and close any open context menu whenever and where ever a click occurred. In that handler, I invoked preventDefault on the event which blocked non scripted actions such as navigation via the Nav bar.
Also, context menus were hidden behind modals. After some trial and error, I discovered that the modals, at least on this page, were given a z-index of 1050. This caused the context menu to render behind the document displayed in the modal view. For now, I have fixed this issue by querying the parents of the right-clicked element until a modal is found, getting its z-index and then setting the z-index of the context menu div to an offset of this.