Significant Improvements to Word Morphers. Improved accuracy of root identification. Improved reflexivity recognition for exception entries. Implemented adverb form recognition. Fixed several bugs and encountered several issues which require discussion.
These revolve in this context around handling multi word (white space, not hyphenated exception entries).
Excepting these, and in light of the introduced improvements, the Word Morpher components, provide reflexive and symmetric relations. That property allows them to, in addition to providing a significant part of the basis for the higher level resuls, aids in validating many other components.

However, if we respond by allowing these morphers to operate directly on phrases, even simply to seek out known exceptional cases, we stand to lose, or at least deminish some of benefits described.