Replaced the few uses of DateTime with DateTimeOffset.
Changed the declarations of all internal typescript modules, replacing the module keyword with the new namespace keyword.

Added a grunt task to remove typescript declaration files included as part of certain bower packages (e.g. angular-ui-router), as they are generally out of date and cause multiple defintion errors in VS, conflicting with the declaration files managed by tsd. Ideally the bower_components folder would simply be excluded from the VS project, but as per https://github.com/aspnet/Tooling/issues/94 this is not yet supported and VS 2015 does not ignore bower_components by default but the pre-release versions did which is why this became a problem.

Upgraded to TypeScript 1.5.3
Updated Angular to 1.4.4.